moved get_data_by_id from extract to connector
